+++ AppCs/AppCs/services/interfaces/AlgorithmInterface.cs
@@ -1,5 +1,5 @@
 namespace services.interfaces;
 public interface AlgorithmInterface{
-        int[][] MultiplyMatrices(int[][] matrix1, int[][] matrix2);
+        long[][] MultiplyMatrices(long[][] matrix1, long[][] matrix2);
 
 }
